References

Contents

References

Contents

Index

Index

CarnotCycles.AmbientNode

CarnotCycles.Compute_cycle_error

CarnotCycles.Compute_cycle_error

CarnotCycles.CoolPropGasPhaseCheck

CarnotCycles.CoolPropLiquidPhaseCheck

CarnotCycles.CoolantPort

CarnotCycles.HeatPort

CarnotCycles.IsentropicCompression

CarnotCycles.IsentropicCompressionClapeyron

CarnotCycles.IsentropicExpansion

CarnotCycles.IsentropicExpansionClapeyron

CarnotCycles.IsochoricCompression

CarnotCycles.IsochoricCompressionClapeyron

CarnotCycles.IsochoricExpansion

CarnotCycles.IsochoricExpansionClapeyron

CarnotCycles.IsothermalCompression

CarnotCycles.IsothermalCompressionClapeyron

CarnotCycles.IsothermalExpansion

CarnotCycles.IsothermalExpansionClapeyron

CarnotCycles.MassSink

CarnotCycles.MassSource

CarnotCycles.PT_IsentropicExpansionClapeyron

CarnotCycles.PowerPort

CarnotCycles.RefPort

CarnotCycles.StoragePort

CarnotCycles.load_fluid

CarnotCycles.load_fluid

CarnotCycles.mass_to_moles

CarnotCycles.AmbientNode — Method

Ambient temperature node.

CarnotCycles.Compute_cycle_error — Method

Compute_cycle_error(p_source,h_source,p_sink,h_sink;reltol = 1e-8) * Computes cycle end point mismatch in states - between sink and source.

returns `nothing`

Shows error incase of mismatch

CarnotCycles.Compute_cycle_error — Method

Compute_cycle_error(sol::ODESolution,system::Vector{ODESystem};reltol = 1e-8)

CarnotCycles.CoolPropGasPhaseCheck — Method

CoolPropGasPhaseCheck Checks if state point is Gas. Passes through an assert statement.

CarnotCycles.CoolPropLiquidPhaseCheck — Method

CoolPropLiquidPhaseCheck Checks if state point is Liquid. Passes through an assert statement.

CarnotCycles.CoolantPort — Method

Makes node for port connections. This node is Pressure,Enthalpy, Mass flow rate and mass fraction of first fluid (incase of Clapyeron Mixture).

CarnotCycles.HeatPort — Method

HeatPort: Variables are Q, tin, Tout

CarnotCycles.IsentropicCompression — Method

IsentropicCompression(πc, h_in, p_in,fluid,η)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

η : Isentropic Efficiency

returns : Outlet enthalpy after isentropic compression

CarnotCycles.IsentropicCompressionClapeyron — Method

IsentropicCompressionClapeyron(πc, h_in, p_in,z,fluid,η)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

z : Moles

η : Isentropic Efficiency

returns : Outlet enthalpy after isentropic compression

CarnotCycles.IsentropicExpansion — Method

IsentropicExpansion(πc, h_in, p_in,fluid,η)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

η : Isentropic Efficiency

returns : Outlet enthalpy after isentropic expansion

CarnotCycles.IsentropicExpansionClapeyron — Method

IsentropicExpansionClapeyron(πc, h_in, p_in,z,fluid,η)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

z : Moles

η : Isentropic Efficiency

returns : Outlet enthalpy after isentropic expansion

CarnotCycles.IsochoricCompression — Method

IsochoricCompression(πc, h_in, p_in,fluid)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

Output -> Outlet enthalpy after isochoric compression

CarnotCycles.IsochoricCompressionClapeyron — Method

IsochoricCompressionClapeyron(πc, h_in, p_in,z::Array,fluid::EoSModel)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

z : Moles

fluid: Fluid

Output -> Outlet enthalpy after isochoric compression

CarnotCycles.IsochoricExpansion — Method

IsochoricExpansion(πc, h_in, p_in,fluid)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

Output -> Outlet enthalpy after isochoric expansion

CarnotCycles.IsochoricExpansionClapeyron — Method

IsochoricExpansionClapeyron(πc, h_in, p_in,z::Array,fluid::EoSModel)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

z : Moles

fluid: Fluid

Output -> Outlet enthalpy after isochoric expansion

CarnotCycles.IsothermalCompression — Method

IsothermalCompression(πc, h_in, p_in,fluid)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

Output -> Outlet enthalpy after Isothermal Compression

CarnotCycles.IsothermalCompressionClapeyron — Method

IsothermalCompressionClapeyron(πc, h_in, p_in,z,fluid::EoSModel)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

z : Moles

fluid: Fluid

Output -> Outlet enthalpy after Isothermal compression

CarnotCycles.IsothermalExpansion — Method

IsothermalExpansion(πc, h_in, p_in,fluid)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

Output -> Outlet enthalpy after Isothermal Expansion

CarnotCycles.IsothermalExpansionClapeyron — Method

IsothermalExpansionClapeyron(πc, h_in, p_in,z,fluid::EoSModel)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

z : Moles

fluid: Fluid

Output -> Outlet enthalpy after Isothermal expansion

CarnotCycles.MassSink — Method

MassSink` - Sets the final port input values to the variables. Use as a sink for the simulation.

CarnotCycles.MassSource — Method

MassSource - Initilizes cycle start point. Requires initial enthalpy,pressure and Mass flow rate if CoolProp fluid is used else uses enthalpy,pressure, Mass flow rate, and mass fraction of first fluid if Clapeyron Fluid is used.

CarnotCycles.PT_IsentropicExpansionClapeyron — Method

PT_IsentropicExpansionClapeyron(model::EoSModel,T_in,p_in,z,πc,η)

Arguments:

πc : Pressure Ratio

T_in : Inlet Temperature

p_in : Inlet Pressure

fluid: Fluid

z : Moles

η : Isentropic Efficiency

returns : Outlet Temperature after isentropic expansion

CarnotCycles.PowerPort — Method

CarnotCycles.RefPort — Method

Makes node for port connections. This node is Pressure,Temperature, Mass flow rate and mass fraction of first fluid (incase of Clapyeron Mixture). Use this when the two-phase details of the fluid are not necessary.

CarnotCycles.StoragePort — Method

Storage port that connect the storage HTF to the thermal storage. Contains Temperature and mass flow rate of the HTF.

CarnotCycles.load_fluid — Method

load_fluid(x::AbstractString) - fixes fluid for simulation through components using CoolProp as backend.

CarnotCycles.load_fluid — Method

load_fluid(x::Clapeyron.EoSModel) - fixes fluid for simulation through components using Clapeyron as backend

CarnotCycles.mass_to_moles — Method

mass_to_moles(model::EoSModel,x,mass) : convert mass of fluid to number of moles based on the composition of 1st fluid by mass x