References

Contents

References

Contents

Index

Index

CarnotCycles.AmbientNode

CarnotCycles.Compute_cycle_error

CarnotCycles.Compute_cycle_error

CarnotCycles.CoolPropGasPhaseCheck

CarnotCycles.CoolPropLiquidPhaseCheck

CarnotCycles.CoolantPort

CarnotCycles.HeatPort

CarnotCycles.IsentropicCompression

CarnotCycles.IsentropicCompressionClapeyron

CarnotCycles.IsentropicCompressor

CarnotCycles.IsentropicExpander

CarnotCycles.IsentropicExpansion

CarnotCycles.IsentropicExpansionClapeyron

CarnotCycles.IsobaricHeatSink

CarnotCycles.IsobaricHeatSource

CarnotCycles.IsochoricCompression

CarnotCycles.IsochoricCompressionClapeyron

CarnotCycles.IsochoricCompressor

CarnotCycles.IsochoricExpansion

CarnotCycles.IsochoricExpansionClapeyron

CarnotCycles.IsothermalCompression

CarnotCycles.IsothermalCompressionClapeyron

CarnotCycles.IsothermalCompressor

CarnotCycles.IsothermalExpansion

CarnotCycles.IsothermalExpansionClapeyron

CarnotCycles.MassSink

CarnotCycles.MassSource

CarnotCycles.PT_IsentropicExpansionClapeyron

CarnotCycles.PackedBed

CarnotCycles.PhaseIdentificationNumeric

CarnotCycles.PowerPort

CarnotCycles.Pump

CarnotCycles.RefPort

CarnotCycles.SimpleCondensor

CarnotCycles.SimpleCondensorGlide

CarnotCycles.SimpleEvaporator

CarnotCycles.SimpleEvaporatorGlide

CarnotCycles.StoragePort

CarnotCycles.Valve

CarnotCycles.is_feasible_condensor

CarnotCycles.is_feasible_evaporator

CarnotCycles.load_fluid

CarnotCycles.load_fluid

CarnotCycles.mass_to_moles

CarnotCycles.show_all_states

CarnotCycles.AmbientNode — Method

Ambient temperature node.

CarnotCycles.Compute_cycle_error — Method

Compute_cycle_error(p_source,h_source,p_sink,h_sink;reltol = 1e-8) * Computes cycle end point mismatch in states - between sink and source.

returns `nothing`

Shows error incase of mismatch

CarnotCycles.Compute_cycle_error — Method

Compute_cycle_error(sol::ODESolution,system::Vector{ODESystem};reltol = 1e-8)

CarnotCycles.CoolPropGasPhaseCheck — Method

CoolPropGasPhaseCheck Checks if state point is Gas. Passes through an assert statement.

CarnotCycles.CoolPropLiquidPhaseCheck — Method

CoolPropLiquidPhaseCheck Checks if state point is Liquid. Passes through an assert statement.

CarnotCycles.CoolantPort — Method

Makes node for port connections. This node is Pressure,Enthalpy, Mass flow rate and mass fraction of first fluid (incase of Clapyeron Mixture).

CarnotCycles.HeatPort — Method

HeatPort: Variables are Q, tin, Tout

CarnotCycles.IsentropicCompression — Method

IsentropicCompression(πc, h_in, p_in,fluid,η)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

η : Isentropic Efficiency

returns : Outlet enthalpy after isentropic compression

CarnotCycles.IsentropicCompressionClapeyron — Method

IsentropicCompressionClapeyron(πc, h_in, p_in,z,fluid,η)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

z : Moles

η : Isentropic Efficiency

returns : Outlet enthalpy after isentropic compression

CarnotCycles.IsentropicCompressor — Method

IsentropicCompressor(;name,fluid=set_fluid)

A compressor with isentropic Effeciency and pressure ratio as a parameter is chosen.

CarnotCycles.IsentropicExpander — Method

IsentropicExpander(;name,fluid=set_fluid)

A expander with isentropic Effeciency and pressure ratio as a parameter is chosen.

CarnotCycles.IsentropicExpansion — Method

IsentropicExpansion(πc, h_in, p_in,fluid,η)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

η : Isentropic Efficiency

returns : Outlet enthalpy after isentropic expansion

CarnotCycles.IsentropicExpansionClapeyron — Method

IsentropicExpansionClapeyron(πc, h_in, p_in,z,fluid,η)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

z : Moles

η : Isentropic Efficiency

returns : Outlet enthalpy after isentropic expansion

CarnotCycles.IsobaricHeatSink — Method

IsobaricHeatSink(;name,fluid = set_fluid)

A heat sink independent of temperature and no pressure drop

Parameters:

Q_dot : Total heat supplied

Local Variables:

P : Power

s_in : Inlet Entropy

p_in : Inlet Pressure

T_in : Inlet Temperature

h_in : Inlet Enthalpy

ρ_in : Inlet Density

s_out : Outlet Entropy

p_out : Outlet Pressure

T_out : Outlet Temperature

h_out : Outlet Enthalpy

ρ_out : Outlet Density

Port Variables:

inport : p and h

outport : p and h

CarnotCycles.IsobaricHeatSource — Method

IsobaricHeatSource(;name,fluid = set_fluid)

A heat source independent of temperature and no pressure drop

Parameters:

Q_dot : Total heat supplied

Local Variables:

P : Power

s_in : Inlet Entropy

p_in : Inlet Pressure

T_in : Inlet Temperature

h_in : Inlet Enthalpy

ρ_in : Inlet Density

s_out : Outlet Entropy

p_out : Outlet Pressure

T_out : Outlet Temperature

h_out : Outlet Enthalpy

ρ_out : Outlet Density

Port Variables:

inport : p and h

outport : p and h

CarnotCycles.IsochoricCompression — Method

IsochoricCompression(πc, h_in, p_in,fluid)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

Output -> Outlet enthalpy after isochoric compression

CarnotCycles.IsochoricCompressionClapeyron — Method

IsochoricCompressionClapeyron(πc, h_in, p_in,z::Array,fluid::EoSModel)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

z : Moles

fluid: Fluid

Output -> Outlet enthalpy after isochoric compression

CarnotCycles.IsochoricCompressor — Method

IsochoricCompressor(;name,fluid = set_fluid)

A compressor with pressure ratio as a parameter is chosen.

CarnotCycles.IsochoricExpansion — Method

IsochoricExpansion(πc, h_in, p_in,fluid)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

Output -> Outlet enthalpy after isochoric expansion

CarnotCycles.IsochoricExpansionClapeyron — Method

IsochoricExpansionClapeyron(πc, h_in, p_in,z::Array,fluid::EoSModel)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

z : Moles

fluid: Fluid

Output -> Outlet enthalpy after isochoric expansion

CarnotCycles.IsothermalCompression — Method

IsothermalCompression(πc, h_in, p_in,fluid)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

Output -> Outlet enthalpy after Isothermal Compression

CarnotCycles.IsothermalCompressionClapeyron — Method

IsothermalCompressionClapeyron(πc, h_in, p_in,z,fluid::EoSModel)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

z : Moles

fluid: Fluid

Output -> Outlet enthalpy after Isothermal compression

CarnotCycles.IsothermalCompressor — Method

IsothermalCompressor(;name,fluid = set_fluid)

A compressor with pressure ratio as a parameter is chosen.

CarnotCycles.IsothermalExpansion — Method

IsothermalExpansion(πc, h_in, p_in,fluid)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

fluid: Fluid

Output -> Outlet enthalpy after Isothermal Expansion

CarnotCycles.IsothermalExpansionClapeyron — Method

IsothermalExpansionClapeyron(πc, h_in, p_in,z,fluid::EoSModel)

Arguments:

πc : Pressure Ratio

h_in : Inlet Enthalpy

p_in : Inlet Pressure

z : Moles

fluid: Fluid

Output -> Outlet enthalpy after Isothermal expansion

CarnotCycles.MassSink — Method

MassSink` - Sets the final port input values to the variables. Use as a sink for the simulation.

CarnotCycles.MassSource — Method

MassSource - Initilizes cycle start point. Requires initial enthalpy,pressure and Mass flow rate if CoolProp fluid is used else uses enthalpy,pressure, Mass flow rate, and mass fraction of first fluid if Clapeyron Fluid is used.

CarnotCycles.PT_IsentropicExpansionClapeyron — Method

PT_IsentropicExpansionClapeyron(model::EoSModel,T_in,p_in,z,πc,η)

Arguments:

πc : Pressure Ratio

T_in : Inlet Temperature

p_in : Inlet Pressure

fluid: Fluid

z : Moles

η : Isentropic Efficiency

returns : Outlet Temperature after isentropic expansion

CarnotCycles.PackedBed — Method

A simple Schuman Packed Bed model.

CarnotCycles.PhaseIdentificationNumeric — Method

PhaseIdentificationNumeric(model::EoSModel,p,h,z)

For identification of phase of the fluid when using Clapeyron.

returns 0 for liquid, 1 for vapour, 2 for Two phase

CarnotCycles.PowerPort — Method

CarnotCycles.Pump — Method

Pump(;name,fluid = set_fluid)

A pump with isentropic Efficiency and pressure ratio as parameter is chosen. Ensure that inlet to the pump is liquid by checking the internal variable LiquidPhase.

CarnotCycles.RefPort — Method

Makes node for port connections. This node is Pressure,Temperature, Mass flow rate and mass fraction of first fluid (incase of Clapyeron Mixture). Use this when the two-phase details of the fluid are not necessary.

CarnotCycles.SimpleCondensor — Method

SimpleCondensor(;name,fluid=set_fluid)

Condenses the fluid to ΔT_sc below the saturation point. If the fluid is above the critical point then cools it to ΔT_sc below the critical temperature. No pressure drop is considered.

CarnotCycles.SimpleCondensorGlide — Method

A simple condensor where the HTF inlet and outlet temperature is passed as a parameter. Has a variable is_feas which checks if the fluid passed through violates temperature profile condition or not.

CarnotCycles.SimpleEvaporator — Method

SimpleCondensor(;name,fluid=set_fluid)

Evaporates the fluid to ΔT_sh above the saturation point. If the fluid is above the critical point then cools it to ΔT_sh above the critical temperature. No pressure drop is considered.

CarnotCycles.SimpleEvaporatorGlide — Method

A simple evaporator where the HTF inlet and outlet temperature is passed as a parameter. Has a variable is_feas which checks if the fluid passed through violates temperature profile condition or not.

CarnotCycles.StoragePort — Method

Storage port that connect the storage HTF to the thermal storage. Contains Temperature and mass flow rate of the HTF.

CarnotCycles.Valve — Method

Valve(;name,fluid)

Isenthalpic Valve with pressure ratio as a parameter.

CarnotCycles.is_feasible_condensor — Method

This checks if the temperature profile inside the condensor violates physics or not.

i.e. It will see for a condensor that if the temperature of the working fluid was always more than the temperature of the heat transfer fluid.
if it is feasible then it will return `true` else `false`

CarnotCycles.is_feasible_evaporator — Method

This checks if the temperature profile inside the evaporator violates physics or not.

i.e. It will see for a evaporator that if the temperature of the working fluid was always less than the temperature of the heat transfer fluid.
if it is feasible then it will return `true` else `false`

CarnotCycles.load_fluid — Method

load_fluid(x::AbstractString) - fixes fluid for simulation through components using CoolProp as backend.

CarnotCycles.load_fluid — Method

load_fluid(x::Clapeyron.EoSModel) - fixes fluid for simulation through components using Clapeyron as backend

CarnotCycles.mass_to_moles — Method

mass_to_moles(model::EoSModel,x,mass) : convert mass of fluid to number of moles based on the composition of 1st fluid by mass x

CarnotCycles.show_all_states — Method

show_all_states(sol::SteadyStateSolution,system::Vector{ODESystem},names::Vector{String})

Arguments:

sol : The solution from the SteadtStateSystem.

system : The vector of components chosen. The first should be MassSource and the last has to be MassSink.

names : The names of the components in the system. Pass it as a vector of strings.

Returns: Prints the state points of the system in a readable format in the terminal.